Move schedule tests into per-module test modules.
